SMILE-1268: refactoring, put isJournal and isBook in PublicationContextConverter
